pyproject.toml
[tool.poetry]
name = "haml"
version = "0.1.0"
description = ""
authors = ["bageljr <aryanagarwal897@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud = "^0.34.0"
Flask = "^2.0.2"
google-cloud-aiplatform = "^1.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
